+++ backend/src/main/java/pw/react/backend/repositories/booking/BookingRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 public interface BookingRepository extends JpaRepository<Booking, Integer>, JpaSpecificationExecutor<Booking> {
     //used to check if Parkly already made a specific booking - kinda optional
-    Optional<Booking> findByUser_UserIdAndProviderExternalBookingId(Integer userId, Long providerExternalBookingId);
+    Optional<Booking> findByUser_UserIdAndProviderExternalBookingId(Integer userId, Integer providerExternalBookingId);
     //finds the latest booking for a given user
     Optional<Booking> findFirstByUser_UserIdOrderByBookingIdDesc(Integer userId);
     Page<Booking> findAllByUser_UserId(Integer userId, Pageable pageable);
